src/index.test.ts
// Assuming you have already translated minithesis functions and classes to TypeScript
import { runTest, integers, lists, Random, TestCase, toNumber, MapDB } from './index'; // Adjust import path as necessary
function sum(arr: number[]): number {
  return arr.reduce((acc, curr) => acc + curr, 0);
}




describe('Minithesis Tests', () => {
  test.each(Array.from({ length: 10 }, (_, i) => i))('finds small list - seed %i', (seed) => {
    //const database = new Map<string, Uint8Array>(); // Assuming a Map is used for the database

    // Define the test function
    const testFn = (testCase: TestCase) => {
      // console.error("test case in test function", testCase);
      //console.log("this:", this);
      const ls = testCase.any(lists(integers(BigInt(0), BigInt(1000 )))).map((i: bigint) => toNumber(i));
      // console.log("example", JSON.stringify(ls));

      // coalnsole.log(ls.length);
      // console.log("after test case in test function", testCase);
      if (sum(ls) > 1000) {
	console.log(ls);
	throw new Error('all bad');
      }

    };

    expect(() => {
      const random = new Random(seed); // Ensure Random class supports seeding
      const database = new MapDB();
      //      console.log('running a test');
      runTest(100, random, database, false)(testFn);
    }).toThrow('all bad');
  });
});
